540djangos/540a-djsel2-system1/540e/django-select2/example/requirements.txt
-e ..
django-redis
django-appconf